requirements.txt
absl-py==1.4.0
beautifulsoup4==4.12.2
brotlipy==0.7.0
cachetools==5.3.0
# certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==39.0.1
cycler==0.11.0
# dassl==0.6.3
einops==0.6.1
entrypoints==0.3
filelock==3.12.0
flake8==3.7.9
fonttools==4.38.0
ftfy==6.1.1
future==0.18.3
gdown==4.7.1
google-auth==2.17.3
google-auth-oauthlib==0.4.6
grad-cam==1.4.8
grpcio==1.54.0
idna==3.4
importlib-metadata==6.6.0
isort==4.3.21
joblib==1.2.0
kiwisolver==1.4.4
littleutils==0.2.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
mccabe==0.6.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.21.5
nvidia-ml-py==11.525.112
nvitop==1.1.2
oauthlib==3.2.2
ogb==1.3.6
opencv-python==4.8.0.74
openTSNE==1.0.0
outdated==0.2.2
packaging==23.1
pandas==1.3.5
Pillow==9.4.0
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.5.0
pycparser==2.21
pyflakes==2.1.1
pyOpenSSL==23.0.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
# setuptools==65.6.3
six==1.16.0
soupsieve==2.4.1
tabulate==0.9.0
tb-nightly==2.12.0a20230113
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
threadpoolctl==3.1.0
# torch==1.12.0
# torchaudio==0.12.0
torchcam==0.3.2
# torchvision==0.13.0
tqdm==4.65.0
ttach==0.0.3
typing_extensions==4.3.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
# wheel==0.38.4
wilds==1.2.2
yacs==0.1.8
yapf==0.29.0
zipp==3.15.0
